refactor: move SearchCard to components/router-search
